Move migrations up, as migration
