+++ migration/longestcommon/lc.go
