+++ provider/testing/resource.go
@@ -11,8 +11,8 @@ import (
 
 	sq "github.com/Masterminds/squirrel"
 	"github.com/cloudquery/cq-provider-sdk/cqproto"
+	"github.com/cloudquery/cq-provider-sdk/migration"
 	"github.com/cloudquery/cq-provider-sdk/provider"
-	"github.com/cloudquery/cq-provider-sdk/provider/migrations"
 	"github.com/cloudquery/cq-provider-sdk/provider/schema"
 	"github.com/cloudquery/cq-provider-sdk/testlog"
 	"github.com/cloudquery/faker/v3"
@@ -61,7 +61,7 @@ func TestResource(t *testing.T, resource ResourceTestCase) {
 	l := testlog.New(t)
 	l.SetLevel(hclog.Debug)
 	resource.Provider.Logger = l
-	tableCreator := migrations.NewTableCreator(l, schema.PostgresDialect{})
+	tableCreator := migration.NewTableCreator(l, schema.PostgresDialect{})
 	if err := tableCreator.CreateTable(context.Background(), conn, resource.Table, nil); err != nil {
 		assert.FailNow(t, fmt.Sprintf("failed to create tables %s", resource.Table.Name), err)
 	}
